+++ lib/rails-settings/cached_settings.rb
@@ -1,28 +1,39 @@
 module RailsSettings
   class CachedSettings < Settings
-    after_commit :rewrite_cache, on: %i(create update)
     def rewrite_cache
-      Rails.cache.write("rails_settings_cached:#{var}", value)
+      Rails.cache.write(cache_key, value)
     end
 
-    after_commit :expire_cache, on: %i(destroy)
     def expire_cache
-      Rails.cache.delete("rails_settings_cached:#{var}")
+      Rails.cache.delete(cache_key)
+    end
+
+    def cache_key
+      self.class.cache_key(var, thing)
     end
 
     class << self
-      def [](var_name)
-        cache_key = "rails_settings_cached:#{var_name}"
-        obj = Rails.cache.read(cache_key)
-        obj = super(var_name) if obj.nil?
+      def cache_key(var_name, scope_object)
+        scope = "rails_settings_cached:"
+        scope << "#{scope_object.class.base_class.to_s}-#{scope_object.id}:" if scope_object
+        scope << "#{var_name}"
+      end
 
-        return @@defaults[var_name.to_s] if obj.nil?
-        obj
+      def [](var_name)
+        value = Rails.cache.fetch(cache_key(var_name, @object)) do
+          super(var_name)
+        end
+
+        if value.nil?
+          @@defaults[var_name.to_s] if value.nil?
+        else
+          value
+        end
       end
 
       def save_default(key, value)
-        return false unless send(key).nil?
-        send("#{key}=", value)
+        return false unless self[key].nil?
+        self[key] = value
       end
     end
   end